requirements.txt
Flask==3.1.1
tensorflow==2.19.0  # works with Python 3.10/3.11
opencv-python
Pillow
numpy
